refactor: split loop and switch semantics
